+++ app/src/main/java/elite/intel/ai/brain/handlers/commands/OpenCommsPanelHandler.java
@@ -22,7 +22,6 @@ public class OpenCommsPanelHandler implements CommandHandler {
 
     @Override public void handle(String action, JsonObject params, String responseText) {
         if (status.isInMainShip() || status.isInFighter()) {
-            GameControllerBus.publish(new GameInputEvent(Bindings.GameCommand.BINDING_TARGET_NEXT_ROUTE_SYSTEM.getGameBinding(), 0));
             navigator.openAndNavigate(StatusFlags.GuiFocus.COMMS_PANEL, CommsPanel.CHAT);
         } else if (status.isOnFoot()) {
             GameControllerBus.publish(new GameInputEvent(BINDING_ON_FOOT_WHEEL.getGameBinding(), 500));