+++ app/src/main/java/elite/intel/ai/brain/handlers/commands/OpenContactsHandler.java
@@ -20,7 +20,6 @@ public class OpenContactsHandler implements CommandHandler {
     @Override
     public void handle(String action, JsonObject params, String responseText) {
         if (status.isInMainShip() || status.isInSrv() || status.isInFighter()) {
-            GameControllerBus.publish(new GameInputEvent(Bindings.GameCommand.BINDING_TARGET_NEXT_ROUTE_SYSTEM.getGameBinding(), 0));
             navigator.openAndNavigate(StatusFlags.GuiFocus.EXTERNAL_PANEL, LeftPanel.CONTACTS);
         } else if (status.isOnFoot()) {
             GameControllerBus.publish(new GameInputEvent(BINDING_ON_FOOT_WHEEL.getGameBinding(), 500));