+++ app/src/main/java/elite/intel/search/edsm/utils/StrongHoldFilter.java
@@ -5,7 +5,7 @@ import elite.intel.search.spansh.starsystems.StarSystemResult;
 
 public class StrongHoldFilter {
 
-    public static boolean skipEnemyStrongHold(StarSystemResult a, StarSystemResult b, String pledgedPower) {
+    public static boolean skipEnemyStarSystemHold(StarSystemResult a, StarSystemResult b, String pledgedPower) {
         if(a == null || b == null) return false;
         StarSystemResult.SystemRecord aRecord = a.getRecord();
         StarSystemResult.SystemRecord bRecord = b.getRecord();
@@ -18,10 +18,7 @@ public class StrongHoldFilter {
                 bRecord.getControllingPower()
         ) && pledgedPower.equalsIgnoreCase(bRecord.getControllingPower());
 
-        boolean aStrongHold = aPower && "stronghold".equalsIgnoreCase(aRecord.getPowerState());
-        boolean bStrongHold = bPower && "stronghold".equalsIgnoreCase(bRecord.getPowerState());
-        if (aStrongHold || bStrongHold) return true;
-        return false;
+        return aPower || bPower;
     }
 
 }